csi_live/requirements.txt
absl-py
jupyter
seaborn
numpy
matplotlib
tqdm
paramiko
dpkt
pyqtgraph
pyqt5
